# backend/docs/openapi.yaml
openapi: '3.0.0'
info:
  title: Contract 961.kz API
  description: |
    API документация для системы электронного подписания документов Contract 961.kz
    
    ## Аутентификация
    Большинство эндпоинтов требуют JWT токен в заголовке Authorization:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    - Общий лимит: 100 запросов/минуту
    - OTP отправка: 3 запроса/15 минут
    - OTP верификация: 10 запросов/5 минут
  version: '1.0.0'
  contact:
    name: Contract 961.kz
    email: support@961.kz
servers:
  - url: http://localhost:3000
    description: Development
  - url: https://api.961.kz
    description: Production
...
paths:
...
  /sign/{token}:
    get:
      tags: [Public Sign]
      summary: Получить информацию о документе для подписания
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Информация о документе
          content:
            application/json:
              schema:
                type: object
                properties:
                  documentName:
                    type: string
                  clientName:
                    type: string
                  status:
                    type: string
        '404':
          description: Ссылка недействительна
  /sign/{token}/send-code:
    post:
      tags: [Public Sign]
      summary: Отправить OTP код
      description: Лимит 3 запроса на 15 минут
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Код отправлен
        '429':
          description: Слишком много запросов
  /sign/{token}/verify:
    post:
      tags: [Public Sign]
      summary: Верифицировать OTP и подписать
      description: Лимит 10 запросов на 5 минут
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  pattern: '^\d{4}$'
      responses:
        '200':
          description: Документ подписан
        '400':
          description: Неверный код
        '429':
          description: Слишком много попыток
  /api/public/verify/{displayId}:
    get:
      tags: [Public Verify]
      summary: Проверить подлинность документа
      parameters:
        - name: displayId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{3}-\d{4}-\d{3}$'
            example: REQ-2026-001
      responses:
        '200':
          description: Результат проверки
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      valid:
                        type: boolean
                        enum: [true]
                      request:
                        type: object
                        properties:
                          displayId:
                            type: string
                          documentName:
                            type: string
                          clientName:
                            type: string
                          status:
                            type: string
                          signedAt:
                            type: string
                            format: date-time
                            nullable: true
                          createdAt:
                            type: string
                            format: date-time
                  - type: object
                    properties:
                      valid:
                        type: boolean
                        enum: [false]
                      error:
                        type: string
  /api/public/verify/{displayId}/document:
    get:
      tags: [Public Verify]
      summary: Скачать подписанный документ
      parameters:
        - name: displayId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PDF документ
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: Документ не найден
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...